+++ src/jsMain/kotlin/dev/inmo/jsuikit/modifiers/UIKitIconnav.kt
@@ -1,0 +1,11 @@
+package dev.inmo.jsuikit.modifiers
+
+sealed class UIKitIconnav(
+    override val classes: Array<String>
+) : UIKitModifier {
+
+    object Vertical : UIKitIconnav(arrayOf("uk-iconnav-vertical"))
+
+    companion object : UIKitIconnav(arrayOf("uk-iconnav"))
+
+}